+++ monero/backends/rpcexceptions.py
@@ -1,0 +1,12 @@
+from .. import exceptions
+
+class RPCError(exceptions.BackendException):
+    pass
+
+
+class Unauthorized(RPCError):
+    pass
+
+
+class MethodNotFound(RPCError):
+    pass
