+++ monero/const.py
@@ -11,3 +11,5 @@ PRIO_UNIMPORTANT = 1
 PRIO_NORMAL = 2
 PRIO_ELEVATED = 3
 PRIO_PRIORITY = 4
+
+RESTRICTED_MAX_TRANSACTIONS = 100
