+++ tests/test_jsonrpcwallet.py
@@ -10,7 +10,7 @@ from monero.wallet import Wallet
 from monero.address import BaseAddress, Address, SubAddress
 from monero.seed import Seed
 from monero.transaction import IncomingPayment, OutgoingPayment, Transaction
-from monero.backends.jsonrpc import JSONRPCWallet
+from monero.backends.rpcwallet import JSONRPCWallet
 
 from .base import JSONTestCase
 
@@ -36,7 +36,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
                 'total_balance': 236153709446071,
                 'total_unlocked_balance': 236153709446071}}
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_seed(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -52,7 +52,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
         self.assertIsInstance(seed, Seed)
         self.assertEqual(seed.phrase, phrase)
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_balance(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -158,7 +158,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
         self.assertIsInstance(subaddr, SubAddress)
         self.assertIsInstance(index, int)
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_incoming_confirmed(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -285,7 +285,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
             self.assertIsInstance(pmt.transaction.fee, Decimal)
             self.assertIsInstance(pmt.transaction.height, int)
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_incoming_confirmed_and_unconfirmed(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -437,7 +437,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
             self.assertIsInstance(pmt.transaction.height, (int, type(None)))
 
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_incoming_unconfirmed(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -639,7 +639,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
         self.assertEqual(pmts[0].amount, Decimal('0.52'))
         self.assertEqual(pmts[1].amount, Decimal('0.0212'))
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_incoming_by_payment_ids(self, mock_post):
         # These queries will use get_bulk_payments RPC method instead of get_transfers
         mock_post.return_value.status_code = 200
@@ -726,7 +726,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
             self.assertIsInstance(pmt.transaction.height, int)
             self.assertIn(pmt.payment_id, ids)
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_outgoing(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -894,7 +894,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
             self.assertIsInstance(pmt.transaction.fee, Decimal)
             self.assertIsInstance(pmt.transaction.height, int)
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_outgoing_confirmed_and_unconfirmed(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -1074,7 +1074,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
             self.assertIsInstance(pmt.transaction.fee, Decimal)
             self.assertIsInstance(pmt.transaction.height, (int, type(None)))
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_outgoing_unconfirmed_only(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -1122,7 +1122,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
         self.assertEqual(len(pmt.destinations), 2)
         self.assertEqual(pmt.destinations[0][1] + pmt.destinations[1][1], pmt.amount)
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_send_transfer(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -1168,7 +1168,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
         self.assertIsInstance(result[0], Transaction)
         self.assertEqual(Decimal('111.086545699972'), result[1])
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_dynamic_ring_size_deprecation(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -1183,7 +1183,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
                         'tx_hash_list': ['401d8021975a0fee16fe84acbfc4d8ba6312e563fa245baba2aac382e787fb60'],
                         'tx_key_list': ['7061d4d939b563a11e344c60938410e2e63ea72c43741fae81b8805cebe5570a']}}
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_export_import_outputs(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
@@ -1196,7 +1196,7 @@ class JSONRPCWalletTestCase(JSONTestCase):
         mock_post.return_value.json.return_value = {u'id': 0, u'jsonrpc': u'2.0', u'result': {u'num_imported': 9}}
         self.assertEqual(self.wallet.import_outputs(outs_hex), 9)
 
-    @patch('monero.backends.jsonrpc.requests.post')
+    @patch('monero.backends.rpcwallet.requests.post')
     def test_export_import_key_images(self, mock_post):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.return_value = self.accounts_result
